requirements.txt
colorama
websockets==12.0
python-dotenv==1.0.0
